Low: tools: Fix overflow in crm_simulate --repeat
Due to the way glib parses integer options and casts them to assign them
to variables, a value of --repeat="-1" wrapped around to 2^32 - 1,
causing a very long loop. Instead, we default it to 0 if a negative
value is passed in.

+++ include/pcmki/pcmki_simulate.h
@@ -37,7 +37,7 @@ extern "C" {
  * \return Standard Pacemaker return code
  */
 int pcmk__profile_dir(pcmk__output_t *out, uint32_t flags, const char *dir,
-                      long long repeat, const char *use_date);
+                      unsigned int repeat, const char *use_date);
 
 /*!
  * \internal
